Adding the craftering links at the bottom of the page (#24)
In PR #23, craftering community project was added.

Also, the "system crafter community" page was added as a ring member in craftering.

This is the solution to make links more obvious. It leverage #+begin_center and a small CSS to format the links at the bottom of the community page.

See the image in attachment.

thanks @shom for the suggestion!

@daviwil what do you think?

Reviewed-on: https://codeberg.org/SystemCrafters/systemcrafters-site/pulls/24
Co-authored-by: Benoit Joly <benoit@benoitj.ca>
Co-committed-by: Benoit Joly <benoit@benoitj.ca>